extract getNumberOfUrlSegments into utils
