+++ packages/utils/src/index.ts
@@ -25,3 +25,4 @@ export * from './envelope';
 export * from './clientreport';
 export * from './ratelimit';
 export * from './baggage';
+export * from './url';
