+++ packages/utils/test/misc.test.ts
@@ -5,7 +5,6 @@ import {
   addExceptionMechanism,
   checkOrSetAlreadyCaught,
   getEventDescription,
-  stripUrlQueryAndFragment,
   uuid4,
 } from '../src/misc';
 
@@ -184,27 +183,6 @@ describe('addContextToFrame', () => {
     expect(frame.pre_context).toEqual(['10: j', '11: k', '12: l', '13: m', '14: n']);
     expect(frame.context_line).toEqual('14: n');
     expect(frame.post_context).toEqual([]);
-  });
-});
-
-describe('stripQueryStringAndFragment', () => {
-  const urlString = 'http://dogs.are.great:1231/yay/';
-  const queryString = '?furry=yes&funny=very';
-  const fragment = '#adoptnotbuy';
-
-  it('strips query string from url', () => {
-    const urlWithQueryString = `${urlString}${queryString}`;
-    expect(stripUrlQueryAndFragment(urlWithQueryString)).toBe(urlString);
-  });
-
-  it('strips fragment from url', () => {
-    const urlWithFragment = `${urlString}${fragment}`;
-    expect(stripUrlQueryAndFragment(urlWithFragment)).toBe(urlString);
-  });
-
-  it('strips query string and fragment from url', () => {
-    const urlWithQueryStringAndFragment = `${urlString}${queryString}${fragment}`;
-    expect(stripUrlQueryAndFragment(urlWithQueryStringAndFragment)).toBe(urlString);
   });
 });
